fix(backend): api response changes and type.ts restructure
